// front-control-horarios/src/app/models/empleado.ts
export interface Empleado {
  Nombre: string;
    AppE: string;
    ApmE: string;
    FechaNac: Date;
    Correo: string;
    Region: string;
    AreaTrabajo: string;
    Contrato: string;
    TurnoActual: {
      HoraInicial: Date;
      HoraFinal: Date;
  },
  HorarioTraining: {
      FechaInicio: Date;
      FechaFin: Date;
      HoraInicial: Date;
      HoraFinal: Date;
  }
}